requirements.txt
streamlit
streamlit-chat
requests
python-dotenv
gspread
google-auth
